requirements.txt
requests==2.32.3
python-telegram-bot>=20.0,<22.0
apscheduler==3.10.4
sqlalchemy>=1.4,<2.0
timezonefinder>=6.0,<7.0
google-genai>=1.0.0
Pillow>=11.0.0
cloudinary>=1.36.0

# dev dependencies
pytest>=8.0.0
pytest-asyncio>=0.23
